I guess overflow: hidden fixes this
I  was encountering a 1 pixel misalignment where sometimes the
highlighted area wouldn't cover the edges of the card, which would be
fixed with a left/right margin of -1. But then, the highlighted area
would clip outside the card and sit on top of the border. Every time I
resized the window, it had a chance of happening, but it wasn't
correlated with any specific window width. So in the end I decided that
I'm gonna have negative margins on both sides, and use overflow: hidden
to ensure that the highlighted area never clips onto the card borders